+++ AdapterUnitTests/BidMachineAdapterTests/BidMachineAdapterTests.swift
@@ -24,10 +24,6 @@ final class BidMachineAdapterTests {
     BidMachineClientFactory.debugClient = FakeBidMachineClient()
   }
 
-  deinit {
-    BidMachineAdapterExtras.isTestMode = false
-  }
-
   @Test("Adapter version validation")
   func adapterVersion_validates() {
     let adapterVersion = BidMachineAdapter.adapterVersion()
@@ -47,7 +43,6 @@ final class BidMachineAdapterTests {
   @Test("Adapter extra validation")
   func adapterExtra_validates() {
     #expect(BidMachineAdapter.networkExtrasClass() == BidMachineAdapterExtras.self)
-    #expect(BidMachineAdapterExtras.isTestMode == false)
   }
 
 }
@@ -64,13 +59,10 @@ final class BidMachineAdapterInitTests {
 
   deinit {
     MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = nil
-    BidMachineAdapterExtras.isTestMode = false
   }
 
-  @Test("Set up succeeds with the test mode off and coppa undefined")
+  @Test("Set up succeeds with coppa undefined")
   func setUp_succeeds_whenTestModeOffAndCOPPAUndefined() {
-    BidMachineAdapterExtras.isTestMode = false
-
     let credentials = AUTKMediationCredentials()
     credentials.settings = ["source_id": "source_id"]
     let serverConfiguration = AUTKMediationServerConfiguration()
@@ -80,33 +72,11 @@ final class BidMachineAdapterInitTests {
       #expect(error == nil)
     }
     #expect(client.sourceId == "source_id")
-    #expect(client.isTestMode == false)
     #expect(client.isCOPPA == nil)
   }
 
-  @Test("Set up succeeds with the test mode on and coppa undefined")
-  func setUp_succeeds_whenTestModeOnAndCOPPAUndefined() async {
-    BidMachineAdapterExtras.isTestMode = true
-
-    let credentials = AUTKMediationCredentials()
-    credentials.settings = ["source_id": "source_id"]
-    let serverConfiguration = AUTKMediationServerConfiguration()
-    serverConfiguration.credentials = [credentials]
-
-    await confirmation("wait for the adpater set up") { setUpCompletion in
-      BidMachineAdapter.setUp(with: serverConfiguration) { error in
-        #expect(error == nil)
-        setUpCompletion()
-      }
-    }
-    #expect(client.sourceId == "source_id")
-    #expect(client.isTestMode == true)
-    #expect(client.isCOPPA == nil)
-  }
-
-  @Test("Set up succeeds with the test mode on and coppa is set to false")
+  @Test("Set up succeeds with coppa is set to false")
   func setUp_succeeds_whenTestModeOnAndCOPPAFalse() async {
-    BidMachineAdapterExtras.isTestMode = true
     MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = false
 
     let credentials = AUTKMediationCredentials()
@@ -121,13 +91,11 @@ final class BidMachineAdapterInitTests {
       }
     }
     #expect(client.sourceId == "source_id")
-    #expect(client.isTestMode == true)
     #expect(client.isCOPPA == false)
   }
 
-  @Test("Set up succeeds with the test mode on and coppa is set to true")
+  @Test("Set up succeeds with coppa is set to true")
   func setUp_succeeds_whenTestModeOnAndCOPPATrue() async {
-    BidMachineAdapterExtras.isTestMode = true
     MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment = true
 
     let credentials = AUTKMediationCredentials()
@@ -142,7 +110,6 @@ final class BidMachineAdapterInitTests {
       }
     }
     #expect(client.sourceId == "source_id")
-    #expect(client.isTestMode == true)
     #expect(client.isCOPPA == true)
   }
 }