+++ adapters/Verve/VerveAdapter/VerveAdapter.swift
@@ -38,7 +38,6 @@ final class VerveAdapter: NSObject, RTBAdapter {
   ) {
     do {
       let sourceId = try Util.appToken(from: configuration)
-      let isTestMode = VerveAdapterExtras.isTestMode
       var isCOPPA: Bool?
       if let COPPA = MobileAds.shared.requestConfiguration.tagForChildDirectedTreatment {
         isCOPPA = COPPA.boolValue
@@ -49,7 +48,7 @@ final class VerveAdapter: NSObject, RTBAdapter {
       }
 
       HybidClientFactory.createClient().initialize(
-        with: sourceId, testMode: isTestMode, COPPA: isCOPPA, TFUA: isTFUA
+        with: sourceId, COPPA: isCOPPA, TFUA: isTFUA
       ) { error in
         completionHandler(error?.toNSError())
       }